--- Sources/muterCore/BuildSystems/Shell/Pipe.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-protocol Pipeable: AnyObject {
-    var fileHandleForReading: FileHandle { get }
-}
-
-extension Pipe: Pipeable {}
